requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.6
MarkupSafe==2.1.2
packaging==23.1
python-dotenv==1.0.0
Werkzeug==2.3.3
